study-pulse-server/requirements.txt
asgiref==3.8.1
Django==5.0.6
python-dotenv==1.0.1
sqlparse==0.5.0
asgiref==3.8.1
attrs==23.2.0
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.1
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
uritemplate==4.1.1
